Refactor: split large files and extract helpers across Rust backend and React frontend
Rust: deduplicate node.rs constructors, remove trivial lock wrappers in core.rs,
extract shared expression parsing and metadata regeneration in declarative_schemas.rs,
add molecule accessor methods to FieldVariant, extract timing log helper in mutation_manager.rs.

React: extract ImageThumbnail, imageUtils, and suggestions from LlmQueryTab;
extract ConfigureAiStep (261 lines) from OnboardingWizard; extract graphUtils and
NodeDetail from WordGraphTab; trim verbose JSDoc to single-line comments in 4 API clients.
