+++ src/server/static-react/src/constants/suggestions.js
@@ -1,0 +1,8 @@
+export const STARTER_SUGGESTIONS = [
+  'What data do I have?',
+  'What taxes did I pay?',
+  'Show my upcoming travel plans',
+  'What medications am I taking?',
+  'Search for recent expenses',
+  'Summarize my contacts',
+]
